% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crrf.R
\name{crrf}
\alias{crrf}
\title{Firth-penalized Fine Gray model}
\usage{
crrf(
  form,
  etype,
  dset,
  firth = TRUE,
  CI = FALSE,
  alpha = 0.05,
  eps = 1e-10,
  CI.print = F
)
}
\arguments{
\item{form}{A formula with the survival model, with usual survival package syntax for the time to event data: Surv(time, event) notation. Values must be column names in dset.}

\item{etype}{A character to indicate event type of interest, as defined by the event indicator in dset.}

\item{dset}{A data.frame including the survival variables defined in the formula and any additional covariates to include in the model.}

\item{firth}{A Boolean of whether to fit with Firth penalty. Default is TRUE.}

\item{CI}{A Boolean of whether or not to compute the (penalized) partial profile likelihood confidence intervals. Default is FALSE.}

\item{alpha}{A numeric of significance level. Default is 0.05.}

\item{eps}{A numeric of threshold to begin confidence interval calculations. Default is eps=1e-10. Not recommended to decrease, as this may lead to errors in calculating the Firth Penalty.}

\item{CI.print}{A Boolean of whether to print the CI calculations for each variable while calculating to monitor progress. Default is FALSE.}
}
\value{
A crrf object, with a list of output, including a table with estimates and CI intervals if calculated and the output from nlminb optimization.
}
\description{
Firth-penalized Fine Gray model
}
\examples{
set.seed(12345)
n=100
tm=rexp(n)
ev=factor(rbinom(n,2,0.3))
x=rnorm(n)
y=rbinom(n,1,0.5)
u=runif(n)
dset=cbind.data.frame(tm=tm,ev=ev,x=x,y=y,u=u)
fgm.res=crrf(Surv(tm,ev)~x+y+u,etype="1",dset,firth=TRUE,CI=FALSE)
}
